chore: introduced modular subtitler design
