+++ src/subtitle_tool/ai/__init__.py
@@ -1,0 +1,6 @@
+from subtitle_tool.ai.base import Subtitler
+from subtitle_tool.ai.factory import AISubtitler
+from subtitle_tool.ai.gemini import GeminiSubtitler
+from subtitle_tool.ai.metrics import OperationMetrics
+
+__all__ = ["AISubtitler", "GeminiSubtitler", "OperationMetrics", "Subtitler"]
